Make the job limit proportional to memory, fix linux build

+++ src/util.h
@@ -104,8 +104,8 @@ double GetLoadAverage();
 
 //Swapped memory is very likely caused by the compiler process running out of ram,
 //as such we consider swapped memory as used memory, therefore negative values are to be expected
-/// @return the Available memory on the machine. If no supported it will return std::numeric_limits<double>::max()
-double GetFreeMemory();
+/// @return the Available memory on the machine. If no supported it will return LONG_MAX
+long GetFreeMemory();
 
 /// a wrapper for getcwd()
 std::string GetWorkingDirectory();
